+++ lib/js/index.d.ts
@@ -1,4 +1,4 @@
-import { Config, LazyImage } from './types';
-declare const _default: (selector?: NodeListOf<LazyImage> | string, config?: Config) => void;
+import { Config, InputTypes } from './types';
+declare const _default: (selector?: InputTypes, config?: Config) => void;
 export default _default;
 //# sourceMappingURL=index.d.ts.map