+++ lib/js/types.d.ts
@@ -1,4 +1,5 @@
 export declare type LazyImage = HTMLDivElement | HTMLImageElement;
+export declare type InputTypes = NodeListOf<LazyImage> | LazyImage[] | string;
 export interface Config extends IntersectionObserverInit {
 }
 //# sourceMappingURL=types.d.ts.map